feat: [ANDROID] no op lib
