+++ FloconAndroid/.idea/gradle.xml
@@ -12,6 +12,7 @@
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/core" />
+            <option value="$PROJECT_DIR$/flocon-base" />
             <option value="$PROJECT_DIR$/grpc-interceptor" />
             <option value="$PROJECT_DIR$/okhttp-interceptor" />
           </set>
