+++ FloconAndroid/core/src/main/java/io/github/openflocon/flocon/plugins/analytics/mapper/AnalyticsItemsMapper.kt
@@ -3,24 +3,14 @@ package io.github.openflocon.flocon.plugins.analytics.model
 import org.json.JSONArray
 import org.json.JSONObject
 
-data class AnalyticsItem(
-    val id: String,
-    val analyticsTableId: String,
-    val eventName: String,
-    val createdAt: Long,
-    val properties: List<AnalyticsPropertiesConfig>,
-) {
-    companion object {
-        fun listToJson(items: Collection<AnalyticsItem>) : JSONArray {
-            val array = JSONArray()
+fun analyticsItemsToJson(items: Collection<AnalyticsItem>) : JSONArray {
+    val array = JSONArray()
 
-            items.forEach {
-                array.put(it.toJson())
-            }
-
-            return array
-        }
+    items.forEach {
+        array.put(it.toJson())
     }
+
+    return array
 }
 
 private fun AnalyticsItem.toJson() : JSONObject {
