+++ FloconAndroid/flocon-base/src/main/java/io/github/openflocon/flocon/core/FloconPlugin.kt
