+++ FloconAndroid/flocon-base/src/main/java/io/github/openflocon/flocon/model/FloconMessageFromServer.kt
@@ -1,0 +1,7 @@
+package io.github.openflocon.flocon.model
+
+data class FloconMessageFromServer(
+    val plugin: String,
+    val method: String,
+    val body: String,
+)
