+++ FloconAndroid/flocon-base/src/main/java/io/github/openflocon/flocon/plugins/analytics/FloconAnalyticsPlugin.kt
@@ -1,0 +1,17 @@
+package io.github.openflocon.flocon.plugins.analytics
+
+import io.github.openflocon.flocon.FloconApp
+import io.github.openflocon.flocon.core.FloconPlugin
+import io.github.openflocon.flocon.plugins.analytics.builder.AnalyticsBuilder
+import io.github.openflocon.flocon.plugins.analytics.model.AnalyticsItem
+
+fun FloconApp.analytics(analyticsName: String): AnalyticsBuilder {
+    return AnalyticsBuilder(
+        analyticsTableId = analyticsName,
+        analyticsPlugin = this.client?.analyticsPlugin,
+    )
+}
+
+interface FloconAnalyticsPlugin : FloconPlugin {
+    fun registerAnalytics(analyticsItems: List<AnalyticsItem>)
+}
