+++ FloconAndroid/flocon-base/src/main/java/io/github/openflocon/flocon/plugins/analytics/model/AnalyticsPropertiesConfig.kt
