+++ FloconAndroid/flocon-base/src/main/java/io/github/openflocon/flocon/plugins/analytics/model/TableItem.kt
@@ -1,0 +1,9 @@
+package io.github.openflocon.flocon.plugins.analytics.model
+
+data class AnalyticsItem(
+    val id: String,
+    val analyticsTableId: String,
+    val eventName: String,
+    val createdAt: Long,
+    val properties: List<AnalyticsPropertiesConfig>,
+)
