+++ FloconAndroid/flocon-base/src/main/java/io/github/openflocon/flocon/plugins/dashboard/dsl/PlainTextDsl.kt
