+++ FloconAndroid/flocon-base/src/main/java/io/github/openflocon/flocon/plugins/dashboard/model/config/ElementConfig.kt
