+++ FloconAndroid/flocon-base/src/main/java/io/github/openflocon/flocon/plugins/database/FloconDatabasePlugin.kt
@@ -1,0 +1,6 @@
+package io.github.openflocon.flocon.plugins.database
+
+import io.github.openflocon.flocon.core.FloconPlugin
+
+
+interface FloconDatabasePlugin : FloconPlugin
