+++ FloconAndroid/flocon-base/src/main/java/io/github/openflocon/flocon/plugins/deeplinks/model/Deeplink.kt
