+++ FloconAndroid/flocon-base/src/main/java/io/github/openflocon/flocon/plugins/files/FloconFilesPlugin.kt
@@ -1,0 +1,5 @@
+package io.github.openflocon.flocon.plugins.files
+
+import io.github.openflocon.flocon.core.FloconPlugin
+
+interface FloconFilesPlugin : FloconPlugin
