+++ FloconAndroid/flocon-base/src/main/java/io/github/openflocon/flocon/plugins/sharedprefs/FloconSharedPrefsPlugin.kt
@@ -1,0 +1,5 @@
+package io.github.openflocon.flocon.plugins.SharedPreferences
+
+import io.github.openflocon.flocon.core.FloconPlugin
+
+interface FloconSharedPreferencesPlugin : FloconPlugin
