+++ FloconAndroid/flocon-base/src/main/java/io/github/openflocon/flocon/plugins/tables/model/TableColumnConfig.kt
