+++ FloconAndroid/flocon-base/src/main/java/io/github/openflocon/flocon/plugins/tables/model/TableItem.kt
@@ -1,0 +1,8 @@
+package io.github.openflocon.flocon.plugins.tables.model
+
+data class TableItem(
+    val id: String,
+    val name: String,
+    val createdAt: Long,
+    val columns: List<TableColumnConfig>,
+)
